deduplicate.py not to delete invisible files

+++ python/modules/deduplicate.py
@@ -1,5 +1,5 @@
 # script_name: deduplicate.py
-from python.model.FileSystemNodeModel import Directory
+from python.model.FileSystemNodeModel import Directory, FileSystemCache
 from python.modules.helper_funcs import delete_empty_directories
 
 
@@ -11,14 +11,15 @@ def deduplicate(dir_node: Directory):
     # set to store hashed values of files
     seen = set()
 
-    _deduplicate(dir_node, seen)
+    deleted_files = _deduplicate(dir_node, seen)
 
     # delete empty directories if any after deduplication
     delete_empty_directories(dir_node, dir_node)
 
+    return deleted_files
 
 def _deduplicate(dir_node: Directory, seen: set):
-
+    deleted = []
     # work on a copy of children to safely modify the list during iteration
     for child in dir_node.children[:]:
 
@@ -29,13 +30,22 @@ def _deduplicate(dir_node: Directory, seen: set):
             _deduplicate(child, seen)
 
         # if the child has been seen before, delete it
-        if hashed_value in seen:
+        if hashed_value in seen and not child.is_invisible():
+            deleted.append(child.path)
             child.delete()
 
         # add hash if child is a File Object
         elif hashed_value:
             seen.add(hashed_value)
 
+    print(f"Deleted: {deleted}")
+    return deleted
 
 if __name__ == "__main__":
+    '''
+    c = FileSystemCache()
+    dir_node = Directory('/Users/yachitrasivakumar/Downloads', c)
+    c.save_to_file()
+    deleted = deduplicate(dir_node)
+    '''
     pass
